requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2019.3.9
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.14
django-allauth==0.45.0
django-configurations==2.2
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-decouple==2.1
django-filter==2.4.0
django-heroku==0.3.1
environ==1.0
functions==0.7.0
gcloud-functions-utils==0.13.0
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-pubsub==2.8.0
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
gunicorn==20.1.0
idna==2.8
libcst==0.3.21
mypy-extensions==0.4.3
oauthlib==3.0.1
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.2
protobuf==3.18.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pypi==2.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.4.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.7.2
six==1.16.0
sqlparse==0.2.4
stripe==2.27.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.24.2
utils==1.0.1
whitenoise==5.3.0
